rearrange folder structure according
