+++ src/drawing/data/data.test.ts
@@ -14,9 +14,11 @@
  * limitations under the License.
  */
 
-import { initialize } from "./index";
+import { initialize } from '../../index';
 
-test("circle constructor is mocked", () => {
+test('data object is mocked', () => {
   initialize();
-  expect(new google.maps.Circle(null)).toBeTruthy();
+  expect(new google.maps.Data(null)).toBeTruthy();
+  const map = new google.maps.Map(null);
+  expect(map.data).toBeTruthy();
 });
