+++ src/drawing/data/data.ts
@@ -14,8 +14,8 @@
  * limitations under the License.
  */
 
-import { Feature } from "./data.feature";
-import { MVCObject } from "./mvcobject";
+import { MVCObject } from '../../maps/event/mvcobject';
+import { Feature } from './data.feature';
 
 export class Data extends MVCObject implements google.maps.Data {
   constructor(opt?: google.maps.Data.DataOptions | null) {
