+++ src/drawing/polygons/circle.test.ts
@@ -14,11 +14,9 @@
  * limitations under the License.
  */
 
-import { initialize } from "./index";
+import { initialize } from '../../index';
 
-test("data object is mocked", () => {
+test('circle constructor is mocked', () => {
   initialize();
-  expect(new google.maps.Data(null)).toBeTruthy();
-  const map = new google.maps.Map(null);
-  expect(map.data).toBeTruthy();
+  expect(new google.maps.Circle(null)).toBeTruthy();
 });
