+++ src/maps/maps/visibleregion.ts
@@ -14,9 +14,15 @@
  * limitations under the License.
  */
 
-import { initialize } from "./index";
+import { MVCObject } from '../event/mvcobject';
 
-test("feature object is mocked", () => {
-  initialize();
-  expect(new google.maps.Data.Feature(null)).toBeTruthy();
-});
+export class VisibleRegion
+  extends MVCObject
+  implements google.maps.VisibleRegion
+{
+  public readonly farLeft: google.maps.LatLng;
+  public readonly farRight: google.maps.LatLng;
+  public readonly latLngBounds: google.maps.LatLngBounds;
+  public readonly nearLeft: google.maps.LatLng;
+  public readonly nearRight: google.maps.LatLng;
+}
